feat: add repository interface

+++ internal/api/dtos/wallet.go
@@ -1,8 +1,8 @@
 package dtos
 
 type TransferRequest struct {
-	IdempotencyKey string `json:"idempotencyKey" validate:"required,min=3,max=32"`
-	FromWalletId   string `json:"fromWalletId" validate:"required"`
-	ToWalletId     string `json:"toWalletId" validate:"required"`
-	Amount         int64  `json:"amount" validate:"required,gt=0"`
+	IdempotencyKey string  `json:"idempotencyKey" validate:"required,min=3,max=32"`
+	FromWalletId   string  `json:"fromWalletId" validate:"required"`
+	ToWalletId     string  `json:"toWalletId" validate:"required"`
+	Amount         float64 `json:"amount" validate:"required,gt=0"`
 }
